+++ src/ActionModal/components/ModalPropose.vue
@@ -109,15 +109,15 @@ import {
   between,
   decimal
 } from "vuelidate/lib/validators"
-import num, { uatoms, atoms, SMALLEST } from "../../scripts/num.js"
+import num, { uatoms, atoms, SMALLEST } from "src/scripts/num.js"
 import isEmpty from "lodash.isempty"
 import trim from "lodash.trim"
-import TmField from "common/TmField"
-import TmFormGroup from "common/TmFormGroup"
-import TmFormMsg from "common/TmFormMsg"
-import ActionModal from "common/ActionModal"
+import TmField from "src/components/common/TmField"
+import TmFormGroup from "src/components/common/TmFormGroup"
+import TmFormMsg from "src/components/common/TmFormMsg"
+import ActionModal from "./ActionModal"
 
-import transaction from "src/components/ActionManager/transactionTypes"
+import transaction from "../utils/transactionTypes"
 
 const isValid = type =>
   type === `Text` || type === `ParameterChange` || type === `SoftwareUpgrade`
