+++ src/components/staking/PageValidator.vue
@@ -209,9 +209,9 @@ import num, { atoms, viewDenom, shortDecimals } from "scripts/num"
 import { formatBech32 } from "src/filters"
 import TmBtn from "common/TmBtn"
 import { ratToBigNumber } from "scripts/common"
-import DelegationModal from "staking/DelegationModal"
-import UndelegationModal from "staking/UndelegationModal"
-import ModalWithdrawRewards from "staking/ModalWithdrawRewards"
+import DelegationModal from "src/ActionModal/components/DelegationModal"
+import UndelegationModal from "src/ActionModal/components/UndelegationModal"
+import ModalWithdrawRewards from "src/ActionModal/components/ModalWithdrawRewards"
 import Bech32 from "common/Bech32"
 import TmPage from "common/TmPage"
 import isEmpty from "lodash.isempty"
