+++ src/components/wallet/PageWallet.vue
@@ -42,7 +42,7 @@ import num from "scripts/num"
 import { mapGetters, mapActions } from "vuex"
 import orderBy from "lodash.orderby"
 import LiCoin from "./LiCoin"
-import SendModal from "wallet/SendModal"
+import SendModal from "src/ActionModal/components/SendModal"
 import Bech32 from "common/Bech32"
 import TmPage from "common/TmPage"
 import TmDataMsg from "common/TmDataMsg"
