+++ test/unit/specs/components/ActionModal/components/ActionModal.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils"
 import Vuelidate from "vuelidate"
-import ActionModal from "src/components/common/ActionModal"
+import ActionModal from "src/ActionModal/components/ActionModal"
 import { focusParentLast } from "directives"
 
 const localVue = createLocalVue()
@@ -11,7 +11,7 @@ let mockSimulate = jest.fn(() => 123456)
 let mockSend = jest.fn()
 let mockSetContext = jest.fn()
 
-jest.mock(`src/components/ActionManager/ActionManager.js`, () => {
+jest.mock(`src/ActionModal/utils/ActionManager.js`, () => {
   return jest.fn(() => {
     return {
       setMessage: jest.fn(),
