+++ test/unit/specs/components/ActionModal/components/SendModal.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils"
 import Vuelidate from "vuelidate"
-import SendModal from "src/components/wallet/SendModal"
+import SendModal from "src/ActionModal/components/SendModal"
 
 describe(`SendModal`, () => {
   const localVue = createLocalVue()
